fix: complete unified event API migration and eliminate all compiler warnings
This commit finalizes the event system refactoring by removing the last
instances of the old process_event() API and fixes all compiler warnings
to achieve a completely clean build.

**Changes:**

1. **layer_manager.hh**: Converted to unified event API
   - Replaced process_event() calls with handle_event()
   - Implemented proper ui_event conversion via Backend::create_event()
   - Uses three-phase routing (capture/target/bubble) for mouse events
   - Added required includes (ui_event.hh, event_router.hh, etc.)

2. **Test Backend Implementations**:
   - test_backend.hh: Implemented create_event() to return mouse_event
   - test_canvas_backend.hh: Implemented create_event() to return mouse_event
   - Enables layer_manager tests to work with unified event system

3. **Test Updates for Unified API**:
   - test_layer_manager_events.cc: Updated TrackingElement to use handle_mouse()
   - test_layer_manager_lifetime.cc: Updated test elements to use handle_mouse()
   - test_layer_manager.cc: Marked unused parameter to fix warning

4. **Compiler Warning Fixes** (8 warnings eliminated):
   - test_rule_of_five.cc (5): Added missing .modifiers={} initializers
   - test_scrollbar_visibility.cc (1): Added [[maybe_unused]] attribute
   - demo.hh (1): Added static_cast for sign conversion

5. **Debug Output Removal**:
   - test_scrollbar_visibility.cc: Removed all std::cerr debug statements
   - Cleaned up visual scrollbar detection logging

**Results:**
- ✅ 1217/1217 test cases passing
- ✅ 7044/7044 assertions passing
- ✅ Zero compiler warnings (excluding doctest library)
- ✅ Old event API completely eliminated from codebase

The unified event system is now fully integrated throughout the framework
with no remaining instances of the deprecated process_event(),
handle_mouse_down(), handle_mouse_up(), handle_mouse_enter(), or
handle_mouse_leave() APIs.

+++ examples/demo.hh
@@ -47,7 +47,7 @@ public:
         // Find Norton Blue theme index (default theme, registered first by conio_backend)
         auto norton_it = std::find(m_theme_names.begin(), m_theme_names.end(), "NU8");
         if (norton_it != m_theme_names.end()) {
-            m_current_theme_index = std::distance(m_theme_names.begin(), norton_it);
+            m_current_theme_index = static_cast<std::size_t>(std::distance(m_theme_names.begin(), norton_it));
         } else {
             m_current_theme_index = 0;  // Fallback to first theme if Norton Blue not found
         }
